Implement multiple and restructuring
