+++ kal-derive/src/common/mod.rs
@@ -1,5 +1,5 @@
 pub mod codegen;
 pub mod config;
-pub mod error;
 pub mod doc_string;
+pub mod error;
 pub mod renamer;
