+++ kal/src/command_fragment/try_from_argument_value/impls/option.rs
@@ -1,0 +1,10 @@
+use crate::{CommandArgumentValue, TryFromArgumentValue, TryFromArgumentValueError};
+
+impl<T: TryFromArgumentValue> TryFromArgumentValue for Option<T> {
+    fn try_from_argument_value(
+        value: CommandArgumentValue,
+        strict: bool,
+    ) -> Result<Self, TryFromArgumentValueError> {
+        Ok(T::try_from_argument_value(value, strict).ok())
+    }
+}
